Good Night, maybe! That is all for tonight, I hope.
Updated .gitignore to exclude cached files for Rmd. caching some blocks
to improve performance of knitr during the development.
Need to remove this from the final version made available.
Need to create an issue on github about this.

Have developed the read and combine using the training data.
need to figure out how to remove columns as the next step.

Once that is completed will repeat with the test data set and
compine the two to produce a final dataset.