Updated import.Rmd setup file
Removed the changing of the working directory steps. This doesn't work in Rmd and is not needed in the script. The scrip will run from the directory and file structure will be specified in the readme.md file. Also removed an unneeded str command from the input activity names code block.